* rename windows folder and files to lowercase * rename content when needed as well to lower
TODO: check if Bundle and BundleBuilder are required
