+++ templates/RN/Projects/Default/windows/Param_ProjectToLower.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29215.179
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Param_ProjectName", "Param_ProjectName\Param_ProjectName.csproj", "{F85E6DD1-F8CB-4177-83CC-DA2361C43C91}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Param_ProjectToLower", "Param_ProjectToLower\Param_ProjectToLower.csproj", "{F85E6DD1-F8CB-4177-83CC-DA2361C43C91}"
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Folly", "..\node_modules\react-native-windows\Folly\Folly.vcxproj", "{A990658C-CE31-4BCC-976F-0FC6B1AF693D}"
 EndProject
