+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/.gitignore
