+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Assets/SplashScreen.scale-200.png
