+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Assets/Square150x150Logo.scale-200.png
