+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Assets/Square44x44Logo.scale-200.png
