+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Assets/Square44x44Logo.targetsize-24_altform-unplated.png
