+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Assets/Wide310x150Logo.scale-200.png
