+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/AutolinkedNativeModules.g.props
