+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/Bundle/index.windows.bundle
@@ -1,0 +1,9 @@
+/**
+ *
+ * Be sure to generate this file using the CLI:
+ *
+ * npx react-native bundle --platform windows --entry-file index.js
+ *   --bundle-output windows\Param_ProjectName\Bundle\index.windows.bundle
+ *   --assets-dest windows\Param_ProjectName\Bundle
+ *
+ */
