+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/MainPage.xaml
@@ -1,8 +1,8 @@
 <Page
-    x:Class="Param_ProjectName.MainPage"
+    x:Class="Param_ProjectToLower.MainPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:Param_ProjectName"
+    xmlns:local="using:Param_ProjectToLower"
     xmlns:react="using:Microsoft.ReactNative"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
