+++ templates/RN/Projects/Default/windows/Param_ProjectToLower/MainPage.xaml.cs
@@ -16,7 +16,7 @@ using Windows.UI.Xaml.Navigation;
 
 // The Blank Page item template is documented at https://go.microsoft.com/fwlink/?LinkId=234238
 
-namespace Param_ProjectName
+namespace Param_ProjectToLower
 {
     /// <summary>
     /// An empty page that can be used on its own or navigated to within a Frame.
